The basic idea is present.
Images can be rendered (very crudely) to a string, and the basic
idea of the functions and the pipelining seems to work. Now it
remains to be seen if I can get masking and such to work properly
(signs point to 'HELL YES'), and how feasible it is to implement
backends for multiple types of image files, OpenGL, WxWidgets,
GTK, and all of their friends.

Order of priorities goes:
  * Implement a pretty, color output method
  * Implement all the needed graphics primitives
  * Document the basic framework
  * Implement loading of various image file types
  * Implement writing of various image file types
  * GTK, WxWidgets, OpenGL, etc