Mostly cabalized the library.
It's all nice and fancy and hierarchical now, has a 'pastel.cabal'
file, and such. I'm not yet entirely sure how to hand the code that
depends solely on GTK. Ideally, I'd like to package it as a separate
library or something. Like, have the central core of colors, types,
primitive shapes, and utility functions, and then have a set of
libraries, like 'Graphics.Pastel.Gtk', and 'Graphics.Pastel.PNG'
for importing and exporting.

That's probably how I'll end up doing it, but I would like to avoid
splitting it up into separate files and stuff. Can I do something
with multiple 'library' definitions in the cabal file?