Split off the costly stuff.
It's a hand-rolled function that builds the palette, the palettized
image, and the length, all at once, as it gets folded over the image.

It's also dog-slow. It uses roughly 87% of total processing time in
that function. The program would complete in about half a second if
I could eliminate that slowness.