Holy FUCK it is fast now.
Using a map to store the lookup table, folding over the image data
with the same logic we used before, and serializing to an alist at
the end seems to be a winner. The whole process takes about 0.8s
now, and most of that seems to come from our `polygon` function.
That could probably be reduced quite a bit with some smarter code
to detect points inside and outside the poly.