Implemented WX draw and runner.
It's kinda slow right now, but the WX documentation doesn't really
suggest a faster way of generating the image. I believe the bottleneck
is the fact that we have to call `pixelBufferSetPixel` on each pixel,
instead of writing it directly.

In unrelated news, I decided that I really wasn't using the literate
features of literate Haskell, so I have reverted all files to plain
Haskell. (It's much easier to edit in Vim)